feature/handler feature/query fixed
