+++ src/main/java/ru/netology/Server.java
@@ -6,10 +6,8 @@ import java.net.Socket;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.LocalDateTime;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.Executors;
 
 public class Server {
@@ -17,12 +15,6 @@ public class Server {
             validPaths = List.of("/index.html", "/spring.svg", "/spring.png", "/resources.html", "/styles.css", "/app.js", "/links.html",
             "/forms.html", "/classic.html", "/events.html", "/events.js");
     private final HashMap<List<String>, Handler> handlePaths;
-    private final int bufferLimit = 4096;
-
-    public static final String GET = "GET";
-    public static final String POST = "POST";
-    final List<String> allowedMethods = List.of(GET, POST);
-
 
     public Server() {
         handlePaths = new HashMap<>();
@@ -45,77 +37,12 @@ public class Server {
             final var in = new BufferedInputStream(socket.getInputStream());
             final var out = new BufferedOutputStream(socket.getOutputStream());
 
-            in.mark(bufferLimit);
-            final var buffer = new byte[bufferLimit];
-            final var read = in.read(buffer);
-
-            final var requestLineDelimiter = new byte[]{'\r', '\n'};
-            final var requestLineEnd = indexOf(buffer, requestLineDelimiter, 0, read);
-            if (requestLineEnd == -1) {
+            var request = Request.getInstance(in);
+            if (request == null) {
                 badRequest(out);
                 return;
             }
 
-            // читаем request line
-            final var requestLine = new String(Arrays.copyOf(buffer, requestLineEnd)).split(" ");
-            if (requestLine.length != 3) {
-                badRequest(out);
-                return;
-            }
-
-            final var method = requestLine[0];
-            if (!allowedMethods.contains(method)) {
-                badRequest(out);
-                return;
-            }
-            System.out.println(method);
-
-            var path = requestLine[1];
-            if (!path.startsWith("/")) {
-                badRequest(out);
-                return;
-            }
-            // root redirect to index.html
-            if ("/".equals(path)) {
-                path = "/index.html";
-            }
-            System.out.println(path);
-
-            // ищем заголовки
-            final var headersDelimiter = new byte[]{'\r', '\n', '\r', '\n'};
-            final var headersStart = requestLineEnd + requestLineDelimiter.length;
-            final var headersEnd = indexOf(buffer, headersDelimiter, headersStart, read);
-            if (headersEnd == -1) {
-                badRequest(out);
-                return;
-            }
-
-            // отматываем на начало буфера
-            in.reset();
-            // пропускаем requestLine
-            in.skip(headersStart);
-
-            final var headersBytes = in.readNBytes(headersEnd - headersStart);
-            final var headers = Arrays.asList(new String(headersBytes).split("\r\n"));
-            System.out.println(headers);
-
-            // для GET тела нет
-            String body = null;
-            if (!method.equals(GET)) {
-                in.skip(headersDelimiter.length);
-                // вычитываем Content-Length, чтобы прочитать body
-                final var contentLength = extractHeader(headers, "Content-Length");
-                if (contentLength.isPresent()) {
-                    final var length = Integer.parseInt(contentLength.get());
-                    final var bodyBytes = in.readNBytes(length);
-
-                    body = new String(bodyBytes);
-                    System.out.println(body);
-                }
-            }
-
-            // создание запроса, поиск обработчика
-            var request = new Request(method, path, headers, body);
             Handler handler = handlePaths.get(request.getRequestKey());
             if (handler != null) {
                 handler.handle(request, out);
@@ -161,14 +88,6 @@ public class Server {
         }
     }
 
-    private static Optional<String> extractHeader(List<String> headers, String header) {
-        return headers.stream()
-                .filter(o -> o.startsWith(header))
-                .map(o -> o.substring(o.indexOf(" ")))
-                .map(String::trim)
-                .findFirst();
-    }
-
     private void badRequest(BufferedOutputStream out) throws IOException {
         out.write((
                 "HTTP/1.1 404 Not Found\r\n" +
@@ -177,19 +96,6 @@ public class Server {
                         "\r\n"
         ).getBytes());
         out.flush();
-    }
-
-    private static int indexOf(byte[] array, byte[] target, int start, int max) {
-        outer:
-        for (int i = start; i < max - target.length + 1; i++) {
-            for (int j = 0; j < target.length; j++) {
-                if (array[i + j] != target[j]) {
-                    continue outer;
-                }
-            }
-            return i;
-        }
-        return -1;
     }
 
     public void addHandler(String method, String path, Handler handler) {
